enrollments/app/requirements.txt
fastapi
uvicorn
motor
aioredis
redis
python-dotenv
pydantic
pytest
pytest-asyncio
pytest-dotenv